MULE-12366: Temporary Queues aren't deleted when JMS is used with BTM.
I have seen the comment "do nothing as the close is handled by the session handle".
However, I was looking for in BTM Code, and messageConsumer is never closed.
Even, DualSessionWrapper,  the single class that uses the MessageConsumerWrapper closes it using this method.
Additionally, we are closing the consumer from mule code using reflection using getMessageConsumer().close().